Dossier_review/requirements.txt
fastapi>=0.110.0
beautifulsoup4>=4.12.0
groq>=0.9.0
jinja2>=3.1.0
lxml>=5.0.0
pandas>=2.1.0
PyMuPDF>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
python-pptx>=0.6.23
qdrant-client>=1.7.0
sentence-transformers>=2.7.0
uvicorn>=0.23.0
